refactor: 优化标签ID类型及实验步骤查询逻辑
## 请求类修改
- 将 6 个教师请求类中的 topicTags 字段类型统一修改为 List<Long>
  - CreateTimedQuizProcedureRequest: List<String> -> List<Long>
  - CreateTopicProcedureRequest: List<String> -> List<Long>
  - InsertTimedQuizProcedureRequest: List<String> -> List<Long>
  - InsertTopicProcedureRequest: List<Integer> -> List<Long>
  - UpdateTimedQuizProcedureRequest: List<String> -> List<Long>
  - UpdateTopicProcedureRequest: List<String> -> List<Long>
- 统一使用 Long 类型表示标签 ID，与数据库中的 BIGINT 类型保持一致

## 服务层修改
- StudentExperimentService: 优化实验详情查询
  - 当实验步骤为空时不再抛出异常，改为返回空列表
  - 调整代码结构，仅在有步骤时才计算进度
- StudentProcedureQueryService: 代码简化
  - 将 Collectors.toList() 替换为 Java 16+ 的 .toList()
- TeacherProcedureCreationService: 适配标签类型变更
  - joinTopicTags 方法参数类型改为 List<Long>
  - 使用 StringBuilder 手动拼接，避免类型转换

+++ src/main/java/com/example/demo/pojo/request/teacher/InsertTopicProcedureRequest.java
@@ -48,7 +48,7 @@ public class InsertTopicProcedureRequest {
     /**
      * 标签ID列表（仅在随机抽取时有效），后端自动拼接成"id1,id2"格式
      */
-    private List<Integer> topicTags;
+    private List<Long> topicTags;
 
     /**
      * 题目类型列表（仅在随机抽取时有效），后端自动拼接成"1,2,3"格式
